refactor: refactor the scripts [SF-942]
