+++ utils/constants.ts
@@ -1,6 +1,12 @@
-const Side = {
+export const Side = {
   LEND: 0,
   BORROW: 1,
 };
 
-export { Side };
+export const BASE_CURRENCY_DECIMALS = 8;
+export const CIRCUIT_BREAKER_LIMIT_RANGE = 500;
+export const LIQUIDATION_THRESHOLD_RATE = 12500;
+export const LIQUIDATION_PROTOCOL_FEE_RATE = 200;
+export const LIQUIDATOR_FEE_RATE = 500;
+export const MINIMUM_RELIABLE_AMOUNT = '100000000000';
+export const ORDER_FEE_RATE = 100;
